Split an array along a dimension

split_along_dim(X, .dim) is equivalent to split_on_dim(X, seq_along_dim(X, .dim))

split_on_dim(X, .dim, f = dimnames(X)[[.dim]] %||% seq_along_dim(X, .dim),
  drop = NULL)

split_on_rows(X, f = rownames(X) %||% seq_along_rows(X), drop = NULL)

split_on_cols(X, f = colnames(X) %||% seq_along_cols(X), drop = NULL)

split_along_dim(X, .dim, drop = NULL, .keep_names = TRUE)

split_along_rows(x)

split_along_cols(x)

Arguments

X
an array, or list of arrays. Atomic vectors without a dimension attribute is treated as a 1 dimensions array. Names of list are preserved.

.dim
a scalar integer, specifying which dimension to split along

drop
passed on to [.

.keep_names
Logical. If TRUE then if the dim being split along has dimnames, then the returned list has those names.

Value

A list of arrays, or if a list of arrays was passed in, then a list of lists of arrays.

Examples

X <- array(1:8, c(2,3,4))
X

#> , , 1
#> 
#>      [,1] [,2] [,3]
#> [1,]    1    3    5
#> [2,]    2    4    6
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3]
#> [1,]    7    1    3
#> [2,]    8    2    4
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3]
#> [1,]    5    7    1
#> [2,]    6    8    2
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3]
#> [1,]    3    5    7
#> [2,]    4    6    8
#> 

split_along_dim(X, 2)

#> [[1]]
#>      [,1] [,2] [,3] [,4]
#> [1,]    1    7    5    3
#> [2,]    2    8    6    4
#> 
#> [[2]]
#>      [,1] [,2] [,3] [,4]
#> [1,]    3    1    7    5
#> [2,]    4    2    8    6
#> 
#> [[3]]
#>      [,1] [,2] [,3] [,4]
#> [1,]    5    3    1    7
#> [2,]    6    4    2    8
#> 

split_on_dim(X, 2, c("a", "a", "b"), drop = FALSE)

#> $a
#> , , 1
#> 
#>      [,1] [,2]
#> [1,]    1    3
#> [2,]    2    4
#> 
#> , , 2
#> 
#>      [,1] [,2]
#> [1,]    7    1
#> [2,]    8    2
#> 
#> , , 3
#> 
#>      [,1] [,2]
#> [1,]    5    7
#> [2,]    6    8
#> 
#> , , 4
#> 
#>      [,1] [,2]
#> [1,]    3    5
#> [2,]    4    6
#> 
#> 
#> $b
#> , , 1
#> 
#>      [,1]
#> [1,]    5
#> [2,]    6
#> 
#> , , 2
#> 
#>      [,1]
#> [1,]    3
#> [2,]    4
#> 
#> , , 3
#> 
#>      [,1]
#> [1,]    1
#> [2,]    2
#> 
#> , , 4
#> 
#>      [,1]
#> [1,]    7
#> [2,]    8
#> 
#>