Split an array along a dimension

split_along_dim(X, .dim) is equivalent to split_on_dim(X, seq_along_dim(X, .dim))

split_on_dim(X, .dim, f = dimnames(X)[[.dim]] %||% seq_along_dim(X, .dim),
  drop = NULL, depth = Inf)

split_on_rows(X, f = rownames(X) %||% seq_along_rows(X), drop = NULL,
  depth = Inf)

split_along_dim(X, .dim, drop = NULL, .keep_names = TRUE, depth = Inf)

split_along_rows(X, drop = NULL, .keep_names = TRUE, depth = Inf)

Arguments

X
an array, or list of arrays. Atomic vectors without a dimension attribute is treated as a 1 dimensions array (Meaning, atomic vectors without a dim attribute are only accepted if .dim is 1. Names of list are preserved. If a list of arrays, all the arrays must have the same length of the dimension being split.

.dim
a scalar integer, specifying which dimension to split along

f
Specify how to split the dimension.

character, integer or factor

passed on to base::split(). Must be the same length as the dimention being split.

a list of vectors

Passed on to base::interaction() then base::split(). Each vector in the list must be the same length as the dimention being split.

a scalar integer

used to split into that many groups of equal size

a numeric vector where all( f < 0 )

used to determin the relative proportions of the group being split. sum(f) must be 1. For example c(0.2, 0.2, 0.6) will return approximatly a 20%-20%-60% split.

drop
passed on to [.

depth
Scalar number, how many levels to recurse down. Set this if you want to explicit treat a list as a vector (that is, a one-dimentional array). (You can alternatively set dim attributes with dim<- on the list to prevent recursion)

.keep_names
Logical. If TRUE then if the dim being split along has dimnames, then the returned list has those names.

Value

A list of arrays, or if a list of arrays was passed in, then a list of lists of arrays.

Examples

X <- array(1:8, c(2,3,4))
X

#> , , 1
#> 
#>      [,1] [,2] [,3]
#> [1,]    1    3    5
#> [2,]    2    4    6
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3]
#> [1,]    7    1    3
#> [2,]    8    2    4
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3]
#> [1,]    5    7    1
#> [2,]    6    8    2
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3]
#> [1,]    3    5    7
#> [2,]    4    6    8
#> 

split_along_dim(X, 2)

#> [[1]]
#>      [,1] [,2] [,3] [,4]
#> [1,]    1    7    5    3
#> [2,]    2    8    6    4
#> 
#> [[2]]
#>      [,1] [,2] [,3] [,4]
#> [1,]    3    1    7    5
#> [2,]    4    2    8    6
#> 
#> [[3]]
#>      [,1] [,2] [,3] [,4]
#> [1,]    5    3    1    7
#> [2,]    6    4    2    8
#> 

split_on_dim(X, 2, c("a", "a", "b"), drop = FALSE)

#> $a
#> , , 1
#> 
#>      [,1] [,2]
#> [1,]    1    3
#> [2,]    2    4
#> 
#> , , 2
#> 
#>      [,1] [,2]
#> [1,]    7    1
#> [2,]    8    2
#> 
#> , , 3
#> 
#>      [,1] [,2]
#> [1,]    5    7
#> [2,]    6    8
#> 
#> , , 4
#> 
#>      [,1] [,2]
#> [1,]    3    5
#> [2,]    4    6
#> 
#> 
#> $b
#> , , 1
#> 
#>      [,1]
#> [1,]    5
#> [2,]    6
#> 
#> , , 2
#> 
#>      [,1]
#> [1,]    3
#> [2,]    4
#> 
#> , , 3
#> 
#>      [,1]
#> [1,]    1
#> [2,]    2
#> 
#> , , 4
#> 
#>      [,1]
#> [1,]    7
#> [2,]    8
#> 
#>